refactor(modeling - menuproduct) : 메뉴 패키지 구조 변경
